:truck: start moving sidebar to file based routing
