+++ docs/home/sidebar.ts
@@ -1,0 +1,1 @@
+export const homeSidebar = { text: 'Home', link: '/' }
