+++ docs/learn/sidebar/conceptualSidebar.ts
@@ -1,0 +1,10 @@
+export const conceptualSidebar = {
+  text: '🚧 Everything is an Action',
+  collapsed: true,
+  items: [
+    { text: '🚧 Actions', link: '/learn/conceptual/actions' },
+    { text: '🚧 ActionCreators', link: '/learn/conceptual/actioncreators' },
+    { text: '🚧 ActionHandlers', link: '/learn/conceptual/actionhandlers' },
+    { text: '🚧 ActionListeners', link: '/learn/conceptual/actionlisteners' },
+  ],
+} as const
