+++ docs/learn/sidebar/guidesSidebar.ts
@@ -1,0 +1,20 @@
+export const guidesSidebar = {
+  text: '🚧 Working with EVMts',
+  collapsed: true,
+  items: [
+    {
+      text: '🚧 Working with Listeners',
+      link: '/learn/guides/listeners',
+    },
+    {
+      text: '🚧 Optimistic updates',
+      link: '/learn/guides/optimisticupdates',
+    },
+    {
+      text: '🚧 Solidity scripting',
+      link: '/learn/guides/scripting',
+    },
+    { text: '🚧 Debugging', link: '/learn/guides/debugging' },
+    { text: '🚧 Testing', link: '/learn/guides/testing' },
+  ],
+}
