+++ docs/learn/sidebar/installationSidebar.ts
@@ -1,0 +1,16 @@
+export const installationSidebar = {
+  text: '🚧 Installation',
+  items: [
+    {
+      text: '🚧 Start a new EVMts project',
+      link: '/learn/installation/newproject',
+    },
+    {
+      text: '🚧 Add EVMts to an existing project',
+      link: '/learn/installation/existingproject',
+    },
+    { text: '🚧 Editor setup', link: '/learn/installation/editor' },
+    { text: '🚧 Bundler setup', link: '/learn/installation/bundler' },
+    { text: '🚧 Developer tool', link: '/learn/installation/developertool' },
+  ],
+}
