+++ docs/reference/actions/contracts/sidebar.ts
@@ -1,0 +1,22 @@
+export const contractsSidebar = {
+  text: '🚧 Contract actions',
+  collapsed: true,
+  items: [
+    {
+      text: '🚧 Contract writes',
+      link: '/reference/actions/state/snapshot'
+    },
+    {
+      text: '🚧 Contract reads',
+      link: '/reference/actions/state/snapshot'
+    },
+    {
+      text: '🚧 Events',
+      link: '/reference/actions/state/createfork'
+    },
+    {
+      text: '🚧 Storage',
+      link: '/reference/actions/state/selectfork'
+    },
+  ]
+} as const
