+++ docs/reference/advanced/sidebar.ts
@@ -1,0 +1,14 @@
+export const advancedSidebar = {
+  text: '🚧 Advanced',
+  collapsed: true,
+  items: [
+    {
+      text: '🚧 Building your own custom Actions',
+      link: '/reference/section/name',
+    },
+    {
+      text: '🚧 Contributing to EVMts',
+      link: '/reference/section/name',
+    },
+  ],
+} as const
