+++ docs/reference/cli/sidebar.ts
@@ -1,0 +1,9 @@
+export const cliReferenceSidebar = {
+  text: '🚧 CLI Reference',
+  collapsed: true,
+  items: [
+    { text: '🚧 Create', link: '/reference/section/name' },
+    { text: '🚧 Initialize', link: '/reference/section/name' },
+    { text: '🚧 Generate', link: '/reference/section/name' },
+  ],
+} as const
