+++ docs/reference/coretypes/sidebar.ts
@@ -1,0 +1,22 @@
+export const coreTypesReferenceSidebar = {
+  text: '🚧 Core types',
+  collapsed: true,
+  items: [
+    { text: '🚧 Action', link: '/reference/coretypes/action' },
+    { text: '🚧 ActionCreator', link: '/reference/coretypes/actioncreator' },
+    { text: '🚧 ActionHandler', link: '/reference/coretypes/actionhandler' },
+    {
+      text: '🚧 ActionListener',
+      link: '/reference/coretypes/actionlistener',
+    },
+    { text: '🚧 AddressBook', link: '/reference/coretypes/addressbook' },
+    { text: '🚧 Chains', link: '/reference/coretypes/chains' },
+    {
+      text: '🚧 JsonRpc and JsonWs',
+      link: '/reference/coretypes/jsonrpc',
+    },
+    { text: '🚧 Wallet', link: '/reference/coretypes/wallet' },
+    { text: '🚧 Vm', link: '/reference/coretypes/vm' },
+    { text: '🚧 BlockExplorer', link: '/reference/coretypes/blockexplorer' },
+  ],
+} as const
