+++ docs/reference/usagewithotherlibraries/sidebar.ts
@@ -1,0 +1,11 @@
+export const otherUsagesSidebar = {
+  text: '🚧 Usage with other Ethereum libraries',
+  collapsed: true,
+  items: [
+    { text: '🚧 Ethers', link: '/reference/section/name' },
+    { text: '🚧 Viem', link: '/reference/section/name' },
+    { text: '🚧 Hardhat', link: '/reference/section/name' },
+    { text: '🚧 Foundry', link: '/reference/section/name' },
+    { text: '🚧 MUD', link: '/reference/section/name' },
+  ],
+} as const
