+++ docs/sidebar/common.ts
@@ -1,0 +1,3 @@
+export const homeSidebar = { text: 'Home', link: '/' }
+export const linkToLearnSidebar = { text: 'Learn', link: '/learn/' }
+export const linkToReferenceSidebar = { text: 'Reference', link: '/reference/' }
